+++ packages/ui/client/src/layouts/Layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FormattedMessage, setLocale, useIntl } from 'umi';
+import { FormattedMessage, setLocale, useIntl, addLocale } from 'umi';
 import Helmet from 'react-helmet';
 import moment from 'moment';
 import cls from 'classnames';
@@ -19,12 +19,28 @@ interface ILayoutProps {
 
 const Layout: React.FC<ILayoutProps> = props => {
   const isMini = isMiniUI();
-  const intl = useIntl();
+  const [loaded, setLoaded] = React.useState(false);
   const [theme, setTheme] = React.useState('dark');
 
   const setMomentLocale = (locale: ILocale = getLocale()) => {
     moment.locale(locale === 'zh-CN' ? 'zh-cn' : 'en');
   };
+
+  React.useEffect(() => {
+    const messages = window.g_service.locales.reduce((curr, acc) => {
+      const localeGroup = Object.entries(acc);
+      localeGroup.forEach(group => {
+        const [lang, message] = group;
+        curr[lang] = { ...curr[lang], ...message };
+      });
+      return curr;
+    }, {});
+
+    Object.keys(messages).forEach((locale: string) => {
+      addLocale(locale, messages[locale]);
+    });
+    setLoaded(true);
+  }, []);
 
   React.useLayoutEffect(() => {
     setMomentLocale();
@@ -74,8 +90,8 @@ const Layout: React.FC<ILayoutProps> = props => {
         value={{
           locale,
           theme,
+          loaded,
           isMini,
-          formatMessage: intl.formatMessage,
           currentProject,
           showLogPanel,
           hideLogPanel,
